refactor: class packaging(baseball, entity, constants)
