+++ src/main/java/Constants/Constants.java
@@ -1,4 +1,4 @@
-package baseball;
+package Constants;
 
 public class Constants {
     public static final int MAX_STRIKE = 3;
